+++ internal/repository/database.go
@@ -3,8 +3,6 @@ package repository
 import (
 	"log"
 
-	"github.com/zuzaaa-dev/stawberry/internal/repository/model"
-
 	"github.com/zuzaaa-dev/stawberry/internal/config"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
@@ -16,14 +14,6 @@ func InitDB(cfg *config.Config) *gorm.DB {
 		log.Fatal("Failed to connect to database:", err)
 	}
 
-	// Auto migrate schemas
-	err = db.AutoMigrate(
-		&model.User{},
-		&model.Store{},
-		&model.Product{},
-		&model.Offer{},
-		&model.Notification{},
-	)
 	if err != nil {
 		log.Fatal("Failed to migrate database:", err)
 	}
